chore: version updates, UI stuff
