+++ FullyNoded/BitSense.xcdatamodeld/BitSense.xcdatamodel/contents
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="19206" systemVersion="20G165" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
+<model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="20086" systemVersion="21E258" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
     <entity name="AuthKeys" representedClassName="AuthKeys" syncable="YES" codeGenerationType="class">
         <attribute name="id" optional="YES" attributeType="UUID" usesScalarValueType="NO"/>
         <attribute name="privateKey" optional="YES" attributeType="Binary"/>
@@ -21,11 +21,15 @@
         <attribute name="isActive" optional="YES" attributeType="Boolean" usesScalarValueType="YES"/>
         <attribute name="isJoinMarket" optional="YES" attributeType="Boolean" usesScalarValueType="YES"/>
         <attribute name="isLightning" optional="YES" attributeType="Boolean" usesScalarValueType="YES"/>
+        <attribute name="isNostr" optional="YES" attributeType="Boolean" usesScalarValueType="YES"/>
         <attribute name="label" optional="YES" attributeType="String"/>
         <attribute name="macaroon" optional="YES" attributeType="Binary"/>
+        <attribute name="nostrPrivkey" optional="YES" attributeType="Binary"/>
+        <attribute name="nostrPubkey" optional="YES" attributeType="Binary"/>
         <attribute name="onionAddress" optional="YES" attributeType="Binary"/>
         <attribute name="rpcpassword" optional="YES" attributeType="Binary"/>
         <attribute name="rpcuser" optional="YES" attributeType="Binary"/>
+        <attribute name="subscribeTo" optional="YES" attributeType="Binary"/>
         <attribute name="uncleJim" optional="YES" attributeType="Boolean" usesScalarValueType="YES"/>
     </entity>
     <entity name="Peers" representedClassName="Peers" syncable="YES" codeGenerationType="class">
@@ -101,12 +105,12 @@
     </entity>
     <elements>
         <element name="AuthKeys" positionX="-36" positionY="126" width="128" height="88"/>
-        <element name="NewNodes" positionX="-54" positionY="135" width="128" height="194"/>
+        <element name="JMWallets" positionX="-36" positionY="288" width="128" height="149"/>
+        <element name="NewNodes" positionX="-54" positionY="135" width="128" height="254"/>
         <element name="Peers" positionX="-36" positionY="234" width="128" height="133"/>
         <element name="Signers" positionX="-27" positionY="189" width="128" height="209"/>
         <element name="Transactions" positionX="-36" positionY="279" width="128" height="149"/>
         <element name="Utxos" positionX="-36" positionY="234" width="128" height="238"/>
         <element name="Wallets" positionX="-36" positionY="180" width="128" height="359"/>
-        <element name="JMWallets" positionX="-36" positionY="288" width="128" height="149"/>
     </elements>
 </model>
